Implement auth with sessions and private notes

+++ src/controllers/notesController.js
@@ -14,11 +14,11 @@ export const getAllNotes = async (req, res, next) => {
       perPage = 10,
       tag,
       search,
-      sortBy = '_id',
-      sortOrder = 'asc',
+      // sortBy = '_id',
+      // sortOrder = 'asc',
     } = req.query;
     const skip = (page - 1) * perPage;
-    const notesQuery = Note.find();
+    const notesQuery = Note.find({userId: req.user._id});
 
     if (search) notesQuery.where({ $text: { $search: search } });
 
@@ -29,7 +29,6 @@ export const getAllNotes = async (req, res, next) => {
       notesQuery
         .skip(skip)
         .limit(perPage)
-        .sort({ [sortBy]: sortOrder }),
     ]);
 
     const totalPages = Math.ceil(totalNotes / perPage);
@@ -48,7 +47,7 @@ export const getAllNotes = async (req, res, next) => {
 
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findById(noteId);
+  const note = await Note.findById({_id: noteId, userId: req.user._id});
 
   if (!note) {
     return next(createHttpError(404, 'Note not found'));
@@ -58,13 +57,16 @@ export const getNoteById = async (req, res, next) => {
 };
 
 export const createNote = async (req, res) => {
-  const note = await Note.create(req.body);
+  const note = await Note.create({
+    ...req.body,
+    userId: req.user._id,
+  });
   res.status(201).json(note);
 };
 
 export const deleteNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndDelete({ _id: noteId });
+  const note = await Note.findOneAndDelete({ _id: noteId, userId: req.user._id });
   if (!note) {
     next(createHttpError(404, 'Note not found'));
     return;
@@ -74,9 +76,11 @@ export const deleteNote = async (req, res, next) => {
 
 export const updateNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndUpdate({ _id: noteId }, req.body, {
-    new: true,
-  });
+  const note = await Note.findOneAndUpdate(
+  { _id: noteId, userId: req.user._id },
+  req.body,
+  { new: true }
+);
   if (!note) {
     next(createHttpError(404, 'Note not found'));
     return;
